new changes

+++ Success/BernoulliSwift.swift
@@ -43,7 +43,7 @@ class Bernoulli_Swift
     func Bernoulli (int n: Int, int k: Int, float p: Float) -> Float
     {
         let q: Float = 1 - p;
-        var Bernoulli_var: Float = C(n,total: k) * pow(p,k) * pow(q, n-k); // check for error
+        var Bernoulli_var: Float = C(n,total: k) * powf(p, k) * powf(q, n-k); // check for error
         return Bernoulli_var;
     }
